requirements.txt
torch
transformers
datasets
trl
numpy
matplotlib
pyyaml